LoginManager -> LoginCommand fixed
