+++ AMIG.OS/Kernel/Kernel.cs
@@ -18,7 +18,7 @@ namespace AMIG.OS.Kernel
         //private Helpers helpers = new Helpers(userManagement);
         //private List<string> commandHistory = new List<string>(); // Liste für Befehle
         //private int historyIndex = -1; // Aktuelle Position in der Befehlsliste
-        private User LoggedInUser;
+        //private User LoggedInUser;
         DateTime starttime;
 
         protected override void BeforeRun()
@@ -39,122 +39,11 @@ namespace AMIG.OS.Kernel
             commandHandler = new CommandHandler(userManagement, fileSystemManager);
 
             Console.Clear();
-            Console.WriteLine("\n\t\t\t _____\r\n\t\t\t/     \\\r\n\t_______/_______\\_______\r\n\t\\\\______AMIG.OS______//\n");
-
-            ShowLoginOptions();
-        }
-
-        public void ShowLoginOptions()
-        {
-            //userManagement.DisplayAllUsers(); // nur zum testen
-            Console.WriteLine("Select an option: ");
-            Console.WriteLine("[1] Login");
-            Console.WriteLine("[2] Register");
-                var key = Console.ReadKey(intercept: true);
-                switch (key.Key)
-                {
-                    case ConsoleKey.Escape:
-                        //helpers.AdiosHelper();
-                        break;
-                    case ConsoleKey.D1:
-                    case ConsoleKey.NumPad1:
-                        Login();
-                        break;
-                    case ConsoleKey.D2:
-                    case ConsoleKey.NumPad2:
-                        Register();
-                        ShowLoginOptions();
-                        break;
-                    default:
-                        Console.WriteLine("Invalid option! Please use a Number as an Input");
-                        ShowLoginOptions();
-                        break;
-                }
-        }
-
-        private void Login()
-        {
-            Console.WriteLine("");
-            Console.Write("Username: ");
-            var username = Console.ReadLine();
-            // Prüfen, ob der Benutzername leer oder nur Leerzeichen ist
-            if (string.IsNullOrWhiteSpace(username))
-            {
-                Console.WriteLine("Username cannot be empty. Please try again.");
-                ShowLoginOptions();
-                return;
-            }
-
-            Console.Write("Password: ");
-            var password = ConsoleHelpers.GetPassword();
-
-            // Prüfen, ob das Passwort leer oder nur Leerzeichen ist
-            if (string.IsNullOrWhiteSpace(password))
-            {
-                Console.WriteLine("Password cannot be empty. Please try again.");
-                ShowLoginOptions();
-                return;
-            }
-
-            if (userManagement.Login(username, password))
-            {
-                LoggedInUser = userManagement.GetUser(username);
-                Console.WriteLine($"Der eingeloggte User heißt: {LoggedInUser.Username} mit der Rolle {string.Join(", ", LoggedInUser.Roles.Select(r => r.RoleName))}");
-                //commandHandler.SetStartTime(DateTime.Now); // Startzeit setzen
-                Console.WriteLine("Login successful!");
-                // Systemstart fortsetzen
-            }
-            else
-            {
-                Console.WriteLine("Invalid credentials. Try again.");
-                ShowLoginOptions();
-            }
+            Console.WriteLine("\n\t\t\t _____\r\n\t\t\t/     \\\r\n\t_______/_______\\_______\r\n\t\\\\______AMIG.OS______//\n");           
+            userManagement.loginManager.ShowLoginOptions();
             
         }
 
-        private void Register()
-        {
-            Console.WriteLine("");
-            Console.Write("Username: ");
-            var username = Console.ReadLine();
-            // Prüfen, ob der Benutzername leer oder nur Leerzeichen ist
-            if (string.IsNullOrWhiteSpace(username))
-            {
-                Console.WriteLine("Username cannot be empty. Please try again.");
-                ShowLoginOptions();
-                return;
-            }
-
-            Console.Write("Password: ");
-            var password = ConsoleHelpers.GetPassword();
-
-            // Prüfen, ob das Passwort leer oder nur Leerzeichen ist
-            if (string.IsNullOrWhiteSpace(password))
-            {
-                Console.WriteLine("Password cannot be empty. Please try again.");
-                ShowLoginOptions();
-                return;
-            }
-
-            Console.Write("Choose a role (Admin or Standard): ");
-            var roleInput = Console.ReadLine().ToLower();
-            if (string.IsNullOrWhiteSpace(roleInput))
-            {
-                Console.WriteLine("Role cannot be empty. Please try again.");
-                ShowLoginOptions();
-                return;
-            }
-
-            if (userManagement.Register(username, password, roleInput))
-            {
-                Console.WriteLine("Registration successful! Please log in.");
-            }
-            else
-            {
-                Console.WriteLine("Registration failed. Username may already exist.");
-            }
-            
-        }
 
         //muss in utils oder so
         //public static void ClearCurrentLine()
@@ -170,7 +59,7 @@ namespace AMIG.OS.Kernel
         {
                 Console.Write("> ");
                 var input = Console.ReadLine();
-                commandHandler.ProcessCommand(input, LoggedInUser);
+                commandHandler.ProcessCommand(input, userManagement.loginManager.LoggedInUser);
 
         //Console.Write(helpers.preInput);
         //var currentInput = "";
